requirements.txt
flask
kafka-python
boto3

